requirements.txt
Flask
whisperx
torch
transformers
numpy
werkzeug
